refactor : 로또 폴더로 로또 관련 파일 이동
