+++ src/game.js
@@ -1,12 +1,27 @@
 import { Console } from '@woowacourse/mission-utils';
-import { LOTTO_PRICE, LOTTO_PRIZES } from './constant';
+import { LOTTO_PRICE, LOTTO_PRIZES, LOTTO_MATCH_TEXT } from './constant';
 
+class LottoResult {
+  constructor(matchText, prize) {
+    this.matchText = matchText;
+    this.prize = prize;
+    this.count = 0;
+  }
+
+  incrementCount() {
+    this.count += 1;
+  }
+
+  getTotalPrize() {
+    return this.count * this.prize;
+  }
+}
 const LOTTO_RESULTS = {
-  3: { matchText: LOTTO_MATCH_TEXT.THREE_MATCH, prize: LOTTO_PRIZES.THREE_MATCH, count: 0 },
-  4: { matchText: LOTTO_MATCH_TEXT.FOUR_MATCH, prize: LOTTO_PRIZES.FOUR_MATCH, count: 0 },
-  5: { matchText: LOTTO_MATCH_TEXT.FIVE_MATCH, prize: LOTTO_PRIZES.FIVE_MATCH, count: 0 },
-  '5+bonus': { matchText: LOTTO_MATCH_TEXT.FIVE_MATCH_WITH_BONUS, prize: LOTTO_PRIZES.FIVE_MATCH_WITH_BONUS, count: 0 },
-  6: { matchText: LOTTO_MATCH_TEXT.SIX_MATCH, prize: LOTTO_PRIZES.SIX_MATCH, count: 0 },
+  3: new LottoResult(LOTTO_MATCH_TEXT.THREE_MATCH, LOTTO_PRIZES.THREE_MATCH),
+  4: new LottoResult(LOTTO_MATCH_TEXT.FOUR_MATCH, LOTTO_PRIZES.FOUR_MATCH),
+  5: new LottoResult(LOTTO_MATCH_TEXT.FIVE_MATCH, LOTTO_PRIZES.FIVE_MATCH),
+  '5+bonus': new LottoResult(LOTTO_MATCH_TEXT.FIVE_MATCH_WITH_BONUS, LOTTO_PRIZES.FIVE_MATCH_WITH_BONUS),
+  6: new LottoResult(LOTTO_MATCH_TEXT.SIX_MATCH, LOTTO_PRIZES.SIX_MATCH),
 };
 
 class Game {
@@ -36,7 +51,7 @@ class Game {
 
   #updateLottoResultCount(key) {
     if (key) {
-      LOTTO_RESULTS[key].count += 1;
+      LOTTO_RESULTS[key].incrementCount();
     }
   }
 
@@ -50,7 +65,7 @@ class Game {
 
   #calculateYield() {
     const ticketCount = this.#purchasedLottoNumbersList.length;
-    const totalPrize = Object.values(LOTTO_RESULTS).reduce((total, { prize, count }) => total + prize * count, 0);
+    const totalPrize = Object.values(LOTTO_RESULTS).reduce((total, result) => total + result.getTotalPrize(), 0);
     const yieldRatio = totalPrize / (ticketCount * LOTTO_PRICE);
     const profitability = (yieldRatio * 100).toFixed(1);
     return profitability;
